refactor(web): move the reconciler into its own module
A pure move, ahead of the tests ADR-0032 §6 makes a prerequisite. Nothing in
app.js could be imported without a page: it has no exports, it reads
`document.getElementById("view")` at module scope, and it calls start() on its
last line. So the reconciler — which is where both of this page's real bugs have
been, and neither was found by anything but a person clicking — was unreachable
from a test.

web/dom.js takes el, wanted, keyOf, isTransient, comparable, morph and
morphChildren. Everything there is a function of its arguments, holds no module
state and never reads `view`, which is exactly what makes it importable. show()
and update() stayed behind: they are the same operations bound to the one element
the page draws into, which is the dependency being avoided.

No behaviour changes and no line was retyped. The moved block is byte-identical
to lines 105-137 and 149-276 of the file it came from, and what remains is
byte-identical to the rest plus one import; both were diffed rather than assumed.

**The embed directive was the trap.** It names its files —
`//go:embed index.html app.css app.js` — deliberately, so that nothing lands in
that directory and gets served by accident. A new module is therefore not served
until it is added, and nothing fails at build time when it is missing. The page
would have 404'd on dom.js in the real binary while working perfectly against
web/testdata/fixture.py, which serves the directory from disk. dom.js is now
embedded, verified by reading the FS out of a built binary rather than by
reading the directive, and embed.go says out loud that this is how a page breaks.

Nothing else needed changing: routePage() registers one route per embedded entry
precisely so the served set and the embedded set cannot drift.
